# docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_jamesneb_playback-backend_internal_handlers_dto.LogEntry:
    properties:
      attributes:
        additionalProperties: true
        type: object
      level:
        example: INFO
        type: string
      message:
        example: Order creation started
        type: string
      service:
        example: order-service
        type: string
      span_id:
        example: 789xyz
        type: string
      timestamp:
        example: "2023-01-01T00:00:00Z"
        type: string
      trace_id:
        example: abc123def456
        type: string
    type: object
  github_com_jamesneb_playback-backend_internal_handlers_dto.LogsQueryResponse:
    properties:
      level:
        example: INFO
        type: string
      logs:
        items:
          $ref: '#/definitions/github_com_jamesneb_playback-backend_internal_handlers_dto.LogEntry'
        type: array
      query:
        example: order failed
        type: string
      service:
        example: order-service
        type: string
      time_range:
        $ref: '#/definitions/github_com_jamesneb_playback-backend_internal_handlers_dto.TimeRange'
    type: object
  github_com_jamesneb_playback-backend_internal_handlers_dto.LogsResponse:
    properties:
      received:
        example: 10
        type: integer
      status:
        example: accepted
        type: string
      timestamp:
        example: "2023-01-01T00:00:00Z"
        type: string
    type: object
  github_com_jamesneb_playback-backend_internal_handlers_dto.TimeRange:
    properties:
      from:
        example: "2023-01-01T00:00:00Z"
        type: string
      to:
        example: "2023-01-01T23:59:59Z"
        type: string
    type: object
  github_com_jamesneb_playback-backend_internal_handlers_schema.Attribute:
    properties:
      key:
        type: string
      value:
        $ref: '#/definitions/github_com_jamesneb_playback-backend_internal_handlers_schema.AttributeValue'
    type: object
  github_com_jamesneb_playback-backend_internal_handlers_schema.AttributeValue:
    properties:
      boolValue:
        type: boolean
      doubleValue:
        type: number
      intValue:
        type: integer
      stringValue:
        type: string
    type: object
  github_com_jamesneb_playback-backend_internal_handlers_schema.LogRecord:
    properties:
      attributes:
        items:
          $ref: '#/definitions/github_com_jamesneb_playback-backend_internal_handlers_schema.Attribute'
        type: array
      body:
        $ref: '#/definitions/github_com_jamesneb_playback-backend_internal_handlers_schema.LogRecordBody'
      droppedAttributesCount:
        type: integer
      flags:
        type: integer
      observedTimeUnixNano:
        type: integer
      severityNumber:
        type: integer
      severityText:
        type: string
      spanId:
        type: string
      timeUnixNano:
        type: integer
      traceId:
        type: string
    type: object
  github_com_jamesneb_playback-backend_internal_handlers_schema.LogRecordBody:
    properties:
      stringValue:
        type: string
    type: object
  github_com_jamesneb_playback-backend_internal_handlers_schema.LogsRequest:
    properties:
      resourceLogs:
        items:
          $ref: '#/definitions/github_com_jamesneb_playback-backend_internal_handlers_schema.ResourceLog'
        type: array
    type: object
  github_com_jamesneb_playback-backend_internal_handlers_schema.Resource:
    properties:
      attributes:
        items:
          $ref: '#/definitions/github_com_jamesneb_playback-backend_internal_handlers_schema.Attribute'
        type: array
      droppedAttributesCount:
        type: integer
    type: object
  github_com_jamesneb_playback-backend_internal_handlers_schema.ResourceLog:
    properties:
      resource:
        $ref: '#/definitions/github_com_jamesneb_playback-backend_internal_handlers_schema.Resource'
      schemaUrl:
        type: string
      scopeLogs:
        items:
          $ref: '#/definitions/github_com_jamesneb_playback-backend_internal_handlers_schema.ScopeLog'
        type: array
    type: object
  github_com_jamesneb_playback-backend_internal_handlers_schema.Scope:
    properties:
      attributes:
        items:
          $ref: '#/definitions/github_com_jamesneb_playback-backend_internal_handlers_schema.Attribute'
        type: array
      droppedAttributesCount:
        type: integer
      name:
        type: string
      version:
        type: string
    type: object
  github_com_jamesneb_playback-backend_internal_handlers_schema.ScopeLog:
    properties:
      logRecords:
        items:
          $ref: '#/definitions/github_com_jamesneb_playback-backend_internal_handlers_schema.LogRecord'
        type: array
      scope:
        $ref: '#/definitions/github_com_jamesneb_playback-backend_internal_handlers_schema.Scope'
    type: object
...
  internal_handlers.CreateTraceRequest:
    properties:
      span_id:
        example: def456
        type: string
      tags:
        additionalProperties:
          type: string
        example:
          service: api
          version: "1.0"
        type: object
      timestamp:
        example: "2023-01-01T00:00:00Z"
        type: string
      trace_id:
        example: abc123
        type: string
    required:
    - span_id
    - trace_id
    type: object
...
  internal_handlers.ErrorResponse:
    properties:
      error:
        example: Invalid request
        type: string
      message:
        example: Field validation failed
        type: string
    type: object
...
  internal_handlers.TraceResponse:
    properties:
      created_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      id:
        example: "1"
        type: string
      message:
        example: Successfully processed
        type: string
      service_name:
        example: my-service
        type: string
      status:
        example: accepted
        type: string
      trace_id:
        example: abc123
        type: string
    type: object
...
host: '{{.Host}}'
info:
  contact: {}
  description: Distributed systems event replay backend
  title: Playback Backend API
  version: '{{.Version}}'
paths:
  /api/v1/logs:
    get:
      description: Get log data for analysis
      parameters:
      - description: Service name
        in: query
        name: service
        type: string
      - description: Log level
        in: query
        name: level
        type: string
      - description: Start time (RFC3339)
        in: query
        name: from
        type: string
      - description: End time (RFC3339)
        in: query
        name: to
        type: string
      - description: Search query
        in: query
        name: q
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_jamesneb_playback-backend_internal_handlers_dto.LogsQueryResponse'
      summary: Get logs
      tags:
      - logs
    post:
      consumes:
      - application/json
      description: Receive log data from OpenTelemetry
      parameters:
      - description: Log data
        in: body
        name: logs
        required: true
        schema:
          $ref: '#/definitions/github_com_jamesneb_playback-backend_internal_handlers_schema.LogsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_jamesneb_playback-backend_internal_handlers_dto.LogsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      summary: Receive logs
      tags:
      - logs
...
  /api/v1/traces:
    post:
      consumes:
      - application/json
      description: |-
        Create a new distributed trace
        Receive trace data from OpenTelemetry with optimized processing
      parameters:
      - description: Trace data
        in: body
        name: traces
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.CreateTraceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.TraceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      summary: Receive traces
      tags:
      - traces
  /api/v1/traces/{id}:
    get:
      description: Get a trace by its ID
      parameters:
      - description: Trace ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.TraceResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      summary: Get trace
      tags:
      - traces
...
swagger: "2.0"